Refactor CalorieTracker to use context for calorie calculations and update App component accordingly
