+++ src/context/ActivityContext.tsx
@@ -1,4 +1,4 @@
-import { createContext, Dispatch, ReactNode, useReducer } from 'react';
+import { createContext, Dispatch, ReactNode, useMemo, useReducer } from 'react';
 import {
   ActivityActions,
   activityReducer,
@@ -13,6 +13,9 @@ interface ActivityProviderProps {
 interface ActivityContextProps {
   state: ActivityState;
   dispatch: Dispatch<ActivityActions>;
+  caloriesConsumed: number;
+  caloriesBurned: number;
+  totalCalories: number;
 }
 
 export const ActivityContext = createContext<ActivityContextProps>(null!);
@@ -20,8 +23,41 @@ export const ActivityContext = createContext<ActivityContextProps>(null!);
 export const ActivityProvider = ({ children }: ActivityProviderProps) => {
   const [state, dispatch] = useReducer(activityReducer, initialSate);
 
+  const caloriesConsumed = useMemo(
+    () =>
+      state.activities.reduce(
+        (total, activity) =>
+          activity.category === 1 ? total + activity.calories : total, // category 1 is for food
+        0
+      ),
+    [state.activities]
+  );
+
+  const caloriesBurned = useMemo(
+    () =>
+      state.activities.reduce(
+        (total, activity) =>
+          activity.category === 2 ? total + activity.calories : total, // category 2 is for exercise
+        0
+      ),
+    [state.activities]
+  );
+
+  const totalCalories = useMemo(
+    () => caloriesConsumed - caloriesBurned,
+    [caloriesConsumed, caloriesBurned]
+  );
+
   return (
-    <ActivityContext.Provider value={{ state, dispatch }}>
+    <ActivityContext.Provider
+      value={{
+        state,
+        dispatch,
+        caloriesConsumed,
+        caloriesBurned,
+        totalCalories,
+      }}
+    >
       {children}
     </ActivityContext.Provider>
   );
